+++ lib/fluttermoji_assets/fluttermojimodel.dart
@@ -46,6 +46,22 @@ const String emptySVGIcon = '''
 <svg xmlns="http://www.w3.org/2000/svg"  viewBox="-20 -20 80 80" width="80px" height="80px"><path fill="#dff0fe" stroke="#4788c7" stroke-miterlimit="10" d="M20,1C9.507,1,1,9.507,1,20s8.507,19,19,19s19-8.507,19-19 S30.493,1,20,1z M6,20c0-7.732,6.268-14,14-14c2.963,0,5.706,0.926,7.968,2.496L8.496,27.968C6.926,25.706,6,22.963,6,20z M20,34 c-2.963,0-5.706-0.926-7.968-2.496l19.472-19.472C33.074,14.294,34,17.037,34,20C34,27.732,27.732,34,20,34z"/></svg>
 ''';
 
+const Map<String, int> defaultFluttermojiOptions = {
+  'topType': 24,
+  'accessoriesType': 0,
+  'hairColor': 1,
+  'facialHairType': 0,
+  'facialHairColor': 1,
+  'clotheType': 4,
+  'eyeType': 6,
+  'eyebrowType': 10,
+  'mouthType': 8,
+  'skinColor': 3,
+  'clotheColor': 8,
+  'style': 0,
+  'graphicType': 0
+};
+
 const List<String> FluttermojiStyle = [
   'Transparent',
   'Circle',